% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_eftv_wkly_hrs}
\alias{add_eftv_wkly_hrs}
\title{Add effective weekly hours}
\usage{
add_eftv_wkly_hrs(resources_tb)
}
\arguments{
\item{resources_tb}{Resources (a tibble)}
}
\value{
Resources (a tibble)
}
\description{
add_eftv_wkly_hrs() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add effective weekly hours. The function returns Resources (a tibble).
}
\keyword{internal}
