% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_main_calcs_tb}
\alias{add_main_calcs_tb}
\title{Add main calculations tibble}
\usage{
add_main_calcs_tb(input_data_ls, OOS_buffer_prop_dbl = 0.1)
}
\arguments{
\item{input_data_ls}{Input data (a list)}

\item{OOS_buffer_prop_dbl}{Occasion of service buffer prop (a double vector), Default: 0.1}
}
\value{
Input data (a list)
}
\description{
add_main_calcs_tb() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add main calculations tibble. The function returns Input data (a list).
}
\keyword{internal}
