% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_meets_non_OOS_wkly_hrs_test}
\alias{add_meets_non_OOS_wkly_hrs_test}
\title{Add meets non occasion of service weekly hours test}
\usage{
add_meets_non_OOS_wkly_hrs_test(resources_tb)
}
\arguments{
\item{resources_tb}{Resources (a tibble)}
}
\value{
Resources (a tibble)
}
\description{
add_meets_non_OOS_wkly_hrs_test() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add meets non occasion of service weekly hours test. Function argument resources_tb specifies the object to be updated. The function returns Resources (a tibble).
}
\keyword{internal}
