% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_n_epsds_per_yr}
\alias{add_n_epsds_per_yr}
\title{Add n episodes per yearr}
\usage{
add_n_epsds_per_yr(main_calcs_tb)
}
\arguments{
\item{main_calcs_tb}{Main calcs (a tibble)}
}
\value{
Main calcs (a tibble)
}
\description{
add_n_epsds_per_yr() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add n episodes per yearr. Function argument main_calcs_tb specifies the object to be updated. The function returns Main calcs (a tibble).
}
\keyword{internal}
