% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_n_OOS_per_yr}
\alias{add_n_OOS_per_yr}
\title{Add n occasion of service per year}
\usage{
add_n_OOS_per_yr(main_calcs_tb)
}
\arguments{
\item{main_calcs_tb}{Main calculations (a tibble)}
}
\value{
Main calculations (a tibble)
}
\description{
add_n_OOS_per_yr() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add n occasion of service per year. The function returns Main calculations (a tibble).
}
\keyword{internal}
