% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_OOS_mins_per_yr}
\alias{add_OOS_mins_per_yr}
\title{Add OOS mins per yr}
\usage{
add_OOS_mins_per_yr(main_calcs_tb, OOS_buffer_prop_dbl = 0.1)
}
\arguments{
\item{main_calcs_tb}{Main calcs (a tibble)}

\item{OOS_buffer_prop_dbl}{OOS buffer prop (a double vector), Default: 0.1}
}
\value{
Main calcs (a tibble)
}
\description{
add_OOS_mins_per_yr() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add oos mins per yr. Function argument main_calcs_tb specifies the object to be updated. The function returns Main calcs (a tibble).
}
\keyword{internal}
