% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_resc_occupcy_tb}
\alias{add_resc_occupcy_tb}
\title{Add resource occupancy tibble}
\usage{
add_resc_occupcy_tb(input_data_ls)
}
\arguments{
\item{input_data_ls}{Input data (a list)}
}
\value{
Input data (a list)
}
\description{
add_resc_occupcy_tb() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add resource occupancy tibble. The function returns Input data (a list).
}
\keyword{internal}
