% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_sfcs_to_var_nms}
\alias{add_sfcs_to_var_nms}
\title{Add suffices to variable names}
\usage{
add_sfcs_to_var_nms(var_nms_chr, data_tb)
}
\arguments{
\item{var_nms_chr}{Variable names (a character vector)}

\item{data_tb}{Data (a tibble)}
}
\value{
No return value, called for side effects.
}
\description{
add_sfcs_to_var_nms() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add suffices to variable names. The function is called for its side effects and does not return a value.
}
\keyword{internal}
