% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_sfx_to_var_nm}
\alias{add_sfx_to_var_nm}
\title{Add suffix to variable name}
\usage{
add_sfx_to_var_nm(var_nm_1L_chr, data_tb)
}
\arguments{
\item{var_nm_1L_chr}{Variable name (a character vector of length one)}

\item{data_tb}{Data (a tibble)}
}
\value{
No return value, called for side effects.
}
\description{
add_sfx_to_var_nm() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add suffix to variable name. The function is called for its side effects and does not return a value.
}
\keyword{internal}
