% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimp-package.R, R/pkg_bimp.R
\docType{package}
\name{bimp-package}
\alias{bimp-package}
\title{bimp: Undertake Health Economic Budget Impact Analysis.}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for undertaking budget impact analyses for mental health projects. This highly preliminary development version of the bimp package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors (matthew.hamilton1@monash.edu).

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for undertaking budget impact analyses for mental health projects. This highly preliminary development version of the bimp package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors (matthew.hamilton1@monash.edu).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/bimp/}
  \item \url{https://github.com/ready4-dev/bimp}
  \item \url{https://www.ready4-dev.com/}
}


Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/bimp/}
  \item \url{https://github.com/ready4-dev/bimp}
  \item \url{https://www.ready4-dev.com/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton1@monash.edu} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
}

}
\keyword{internal}
