% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_bind.R
\name{bind_tbs_in_input_ls}
\alias{bind_tbs_in_input_ls}
\title{Bind tibbles in input}
\usage{
bind_tbs_in_input_ls(bc_ls, cf_ls)
}
\arguments{
\item{bc_ls}{Bc (a list)}

\item{cf_ls}{Cf (a list)}
}
\value{
Combined (a list)
}
\description{
bind_tbs_in_input_ls() is a Bind function that binds two instances of a data-frame type class together. Specifically, this function implements an algorithm to bind tibbles in input list. The function returns Combined (a list).
}
\keyword{internal}
