% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_expenditure}
\alias{calculate_expenditure}
\title{Calculate expenditure}
\usage{
calculate_expenditure(
  resources_tb,
  incld_areas_chr = NA_character_,
  area_var_nm_1L_chr = "Recipient_STE_chr",
  n_rescs__var_nm_1L_chr = "Unit_Qty_dbl",
  unit_cost_var_nm_1L_chr = "Unit_Cost_Dollars_dbl"
)
}
\arguments{
\item{resources_tb}{Resources (a tibble)}

\item{incld_areas_chr}{Included areas (a character vector), Default: 'NA'}

\item{area_var_nm_1L_chr}{Area variable name (a character vector of length one), Default: 'Recipient_STE_chr'}

\item{n_rescs__var_nm_1L_chr}{N resources  variable name (a character vector of length one), Default: 'Unit_Qty_dbl'}

\item{unit_cost_var_nm_1L_chr}{Unit cost variable name (a character vector of length one), Default: 'Unit_Cost_Dollars_dbl'}
}
\value{
Expenditure (a double vector of length one)
}
\description{
calculate_expenditure() is a Calculate function that performs a numeric calculation. Specifically, this function implements an algorithm to calculate expenditure. The function returns Expenditure (a double vector of length one).
}
