% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_fns.R
\name{.data}
\alias{.data}
\title{Dot Data function}
\value{
Does not return a value. A pronoun for use in data-masked functions.
}
\description{
Import of \code{\link[rlang:dot-data]{rlang::.data()}}. See rlang package documentation for detailed description.
}
\seealso{
\code{\link[rlang:dot-data]{rlang::.data()}}
}
\keyword{internal}
