% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_fn_type_lup_tb.R
\docType{data}
\name{fn_type_lup_tb}
\alias{fn_type_lup_tb}
\title{Function type lookup table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 46 rows and 6 columns.
}
\usage{
fn_type_lup_tb
}
\description{
A lookup table to find descriptions for different types of functions used within the bimp package suite.
}
\details{
A tibble

\describe{
\item{fn_type_nm_chr}{Function type name (a character vector)}
\item{fn_type_desc_chr}{Function type description (a character vector)}
\item{first_arg_desc_chr}{First argument description (a character vector)}
\item{second_arg_desc_chr}{Second argument description (a character vector)}
\item{is_generic_lgl}{Is generic (a logical vector)}
\item{is_method_lgl}{Is method (a logical vector)}
}
}
\keyword{datasets}
