% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_format.R
\name{format_bound_resc_tb}
\alias{format_bound_resc_tb}
\title{Format bound resource}
\usage{
format_bound_resc_tb(bound_resc_tb, simple_outp_1L_lgl = T)
}
\arguments{
\item{bound_resc_tb}{Bound resource (a tibble)}

\item{simple_outp_1L_lgl}{Simple output (a logical vector of length one), Default: T}
}
\value{
Bound resource (a tibble)
}
\description{
format_bound_resc_tb() is a Format function that binds two instances of a data-frame type class together. Specifically, this function implements an algorithm to format bound resource tibble. The function returns Bound resource (a tibble).
}
\keyword{internal}
