% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_nbr_of_decimals}
\alias{get_nbr_of_decimals}
\title{Get number of decimals}
\usage{
get_nbr_of_decimals(nbr_1L_dbl)
}
\arguments{
\item{nbr_1L_dbl}{Number (a double vector of length one)}
}
\value{
No return value, called for side effects.
}
\description{
get_nbr_of_decimals() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get number of decimals. The function is called for its side effects and does not return a value.
}
\keyword{internal}
