% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_import.R
\name{import_input_data}
\alias{import_input_data}
\title{Import input data}
\usage{
import_input_data(path_to_data_1L_chr)
}
\arguments{
\item{path_to_data_1L_chr}{Path to data (a character vector of length one)}
}
\value{
Input data (a list)
}
\description{
import_input_data() is an Import function that reads a data object in its native format and converts it to an R object. Specifically, this function implements an algorithm to import input data. The function returns Input data (a list).
}
