% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_expenditure_summ_tb}
\alias{make_expenditure_summ_tb}
\title{Make expenditure summ}
\usage{
make_expenditure_summ_tb(
  input_data_ls,
  incld_areas_ls,
  area_var_nm_1L_chr = "Recipient_STE_chr",
  n_ress_var_nm_1L_chr = "Unit_Qty_dbl",
  unit_cost_var_nm_1L_chr = "Unit_Cost_Dollars_dbl"
)
}
\arguments{
\item{input_data_ls}{Input data (a list)}

\item{incld_areas_ls}{Incld areas (a list)}

\item{area_var_nm_1L_chr}{Area var name (a character vector of length one), Default: 'Recipient_STE_chr'}

\item{n_ress_var_nm_1L_chr}{N ress var name (a character vector of length one), Default: 'Unit_Qty_dbl'}

\item{unit_cost_var_nm_1L_chr}{Unit cost var name (a character vector of length one), Default: 'Unit_Cost_Dollars_dbl'}
}
\description{
make_expenditure_summ_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make expenditure summ tibble. The function is called for its side effects and does not return a value.
}
\keyword{internal}
