% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_fns.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\arguments{
\item{lhs, rhs}{An object and a function to apply to it}
}
\description{
Implements: https://github.com/sckott/analogsea/issues/32
and https://github.com/rstudio/ggvis/blob/master/R/pipe.R.
}
\keyword{internal}
