% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_expdr_smry_tb}
\alias{print_expdr_smry_tb}
\title{Print expenditure summary tibble}
\usage{
print_expdr_smry_tb(
  expdr_smry_tb,
  caption_1L_chr = "Predicted Expenditure",
  mkdn_tbl_ref_1L_chr = "tab:exptst",
  output_type_1L_chr = "HTML",
  use_rdocx_1L_lgl = T
)
}
\arguments{
\item{expdr_smry_tb}{Expenditure summary (a tibble)}

\item{caption_1L_chr}{Caption (a character vector of length one), Default: 'Predicted Expenditure'}

\item{mkdn_tbl_ref_1L_chr}{Markdown table reference (a character vector of length one), Default: 'tab:exptst'}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'HTML'}

\item{use_rdocx_1L_lgl}{Use rdocx (a logical vector of length one), Default: T}
}
\description{
print_expdr_smry_tb() is a Print function that prints output to console Specifically, this function implements an algorithm to print expenditure summary tibble. The function is called for its side effects and does not return a value.
}
\keyword{internal}
