% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_expdr_summ_tb}
\alias{print_expdr_summ_tb}
\title{Print expdr summ}
\usage{
print_expdr_summ_tb(
  expdr_summ_tb,
  caption_1L_chr = "Predicted Expenditure",
  mkdn_tbl_ref_1L_chr = "tab:exptst",
  output_type_1L_chr = "HTML"
)
}
\arguments{
\item{expdr_summ_tb}{Expdr summ (a tibble)}

\item{caption_1L_chr}{Caption (a character vector of length one), Default: 'Predicted Expenditure'}

\item{mkdn_tbl_ref_1L_chr}{Mkdn table reference (a character vector of length one), Default: 'tab:exptst'}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'HTML'}
}
\description{
print_expdr_summ_tb() is a Print function that prints output to console Specifically, this function implements an algorithm to print expdr summ tibble. The function is called for its side effects and does not return a value.
}
\keyword{internal}
