% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_remove.R
\name{remove_col_nms_obj_sfcs}
\alias{remove_col_nms_obj_sfcs}
\title{Remove column names object suffices}
\usage{
remove_col_nms_obj_sfcs(
  data_tb,
  complete_cases_1L_lgl = F,
  remove_underscore_1L_lgl = T
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{complete_cases_1L_lgl}{Complete cases (a logical vector of length one), Default: F}

\item{remove_underscore_1L_lgl}{Remove underscore (a logical vector of length one), Default: T}
}
\value{
Data (a tibble)
}
\description{
remove_col_nms_obj_sfcs() is a Remove function that edits an object, removing a specified element or elements. Specifically, this function implements an algorithm to remove column names object suffices. Function argument data_tb specifies the object to be updated. Argument complete_cases_1L_lgl provides the object to be updated. The function returns Data (a tibble).
}
\keyword{internal}
