% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_inp_data_for_incld_intvs}
\alias{transform_inp_data_for_incld_intvs}
\title{Transform input data for included interventions}
\usage{
transform_inp_data_for_incld_intvs(
  input_data_ls,
  incld_intvs_cats_chr = NA_character_,
  incld_intvs_chr = NA_character_,
  intv_uid_var_nm_1L_chr = "Intervention_UID_chr",
  intv_cat_uid_var_nm_1L_chr = "Intervention_Cat_UID_chr"
)
}
\arguments{
\item{input_data_ls}{Input data (a list)}

\item{incld_intvs_cats_chr}{Included interventions cats (a character vector), Default: 'NA'}

\item{incld_intvs_chr}{Included interventions (a character vector), Default: 'NA'}

\item{intv_uid_var_nm_1L_chr}{Intervention unique identifier variable name (a character vector of length one), Default: 'Intervention_UID_chr'}

\item{intv_cat_uid_var_nm_1L_chr}{Intervention cat unique identifier variable name (a character vector of length one), Default: 'Intervention_Cat_UID_chr'}
}
\value{
Input data (a list)
}
\description{
transform_inp_data_for_incld_intvs() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform input data for included interventions. Function argument input_data_ls specifies the object to be updated. Argument incld_intvs_cats_chr provides the object to be updated. The function returns Input data (a list).
}
\keyword{internal}
