% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_inp_data_for_ress_calcs}
\alias{transform_inp_data_for_ress_calcs}
\title{Transform input data for ress calcs}
\usage{
transform_inp_data_for_ress_calcs(input_data_ls)
}
\arguments{
\item{input_data_ls}{Input data (a list)}
}
\value{
Input data (a list)
}
\description{
transform_inp_data_for_ress_calcs() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform input data for ress calcs. Function argument input_data_ls specifies the object to be updated. The function returns Input data (a list).
}
\keyword{internal}
