% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_inp_ls_for_analysis}
\alias{transform_inp_ls_for_analysis}
\title{Transform input list for analysis}
\usage{
transform_inp_ls_for_analysis(input_data_ls, OOS_buffer_prop_dbl = 0.1)
}
\arguments{
\item{input_data_ls}{Input data (a list)}

\item{OOS_buffer_prop_dbl}{OOS buffer prop (a double vector), Default: 0.1}
}
\value{
Transformed input data (a list)
}
\description{
transform_inp_ls_for_analysis() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform input list for analysis. Function argument input_data_ls specifies the object to be updated. Argument OOS_buffer_prop_dbl provides the object to be updated. The function returns Transformed input data (a list).
}
\keyword{internal}
