% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_intv_rcps_per_OOS}
\alias{update_intv_rcps_per_OOS}
\title{Update intervention recipients per occasion of service}
\usage{
update_intv_rcps_per_OOS(main_calcs_tb)
}
\arguments{
\item{main_calcs_tb}{Main calculations (a tibble)}
}
\value{
No return value, called for side effects.
}
\description{
update_intv_rcps_per_OOS() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update intervention recipients per occasion of service. The function is called for its side effects and does not return a value.
}
\keyword{internal}
