% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_intv_recps_per_OOS}
\alias{update_intv_recps_per_OOS}
\title{Update intv recps per OOS}
\usage{
update_intv_recps_per_OOS(main_calcs_tb)
}
\arguments{
\item{main_calcs_tb}{Main calcs (a tibble)}
}
\description{
update_intv_recps_per_OOS() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update intv recps per oos. Function argument main_calcs_tb specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
