% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_resc_occupcy_tb}
\alias{update_resc_occupcy_tb}
\title{Update resource occupancy}
\usage{
update_resc_occupcy_tb(resc_occupcy_tb)
}
\arguments{
\item{resc_occupcy_tb}{Resource occupancy (a tibble)}
}
\value{
Resource occupancy (a tibble)
}
\description{
update_resc_occupcy_tb() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update resource occupancy tibble. Function argument resc_occupcy_tb specifies the object to be updated. The function returns Resource occupancy (a tibble).
}
\keyword{internal}
