% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_max_wkly_OOS_hrs}
\alias{add_max_wkly_OOS_hrs}
\title{Add maximum weekly occasion of service hours}
\usage{
add_max_wkly_OOS_hrs(resources_tb)
}
\arguments{
\item{resources_tb}{Resources (a tibble)}
}
\value{
Resources (a tibble)
}
\description{
add_max_wkly_OOS_hrs() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add maximum weekly occasion of service hours. The function returns Resources (a tibble).
}
\keyword{internal}
