% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_OOS_mins_per_yr}
\alias{add_OOS_mins_per_yr}
\title{Add occasion of service minimums per year}
\usage{
add_OOS_mins_per_yr(main_calcs_tb, OOS_buffer_prop_dbl = 0.1)
}
\arguments{
\item{main_calcs_tb}{Main calculations (a tibble)}

\item{OOS_buffer_prop_dbl}{Occasion of service buffer prop (a double vector), Default: 0.1}
}
\value{
Main calculations (a tibble)
}
\description{
add_OOS_mins_per_yr() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add occasion of service minimums per year. The function returns Main calculations (a tibble).
}
\keyword{internal}
