% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_sfcs_to_var_nms}
\alias{add_sfcs_to_var_nms}
\title{Add suffices to variable names}
\usage{
add_sfcs_to_var_nms(var_nms_chr, data_tb)
}
\arguments{
\item{var_nms_chr}{Variable names (a character vector)}

\item{data_tb}{Data (a tibble)}
}
\description{
add_sfcs_to_var_nms() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add suffices to variable names. Function argument var_nms_chr specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
