% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_sfx_to_var_nm}
\alias{add_sfx_to_var_nm}
\title{Add suffix to var name}
\usage{
add_sfx_to_var_nm(var_nm_1L_chr, data_tb)
}
\arguments{
\item{var_nm_1L_chr}{Var name (a character vector of length one)}

\item{data_tb}{Data (a tibble)}
}
\description{
add_sfx_to_var_nm() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add suffix to var name. Function argument var_nm_1L_chr specifies the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
