% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_bimp.R
\docType{package}
\name{bimp-package}
\alias{bimp}
\alias{bimp-package}
\title{bimp: Undertake Budget Impact Analysis on Projets Developped
with the Ready4 Suite}
\description{
Tools for undertaking budget impact analyses for
mental health projects.  This development version of the bimp package
has been made available as part of the process of testing and
documenting the package. The tools contained in this development
release automate a number of tasks which MODIFY THE DIRECTORY
STRUCTURE OF YOUR LOCAL MACHINE.  Therefore you should only trial this
software if you feel confident that you understand what it does and
have created a sandpit area in which you can safely undertake testing.
If you have any questions, please contact the authors
(matthew.hamilton@orygen.org.au).  The documentation for this package
has been automatically generated by the ready4fun package and is
therefore quite rudimentary. Human authored documentation will follow
in 2021.
}
\details{
To learn more about bimp, start with the vignettes:
\code{browseVignettes(package = "bimp")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/bimp/}
  \item \url{https://github.com/ready4-dev/bimp}
  \item \url{https://ready4-dev.github.io/ready4/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton@orygen.org.au} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
}

}
