% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_bind.R
\name{bind_resource_tbs}
\alias{bind_resource_tbs}
\title{Bind resource tibbles}
\usage{
bind_resource_tbs(res_use_tb, resources_tb)
}
\arguments{
\item{res_use_tb}{Res use (a tibble)}

\item{resources_tb}{Resources (a tibble)}
}
\value{
Joined (a tibble)
}
\description{
bind_resource_tbs() is a Bind function that binds two instances of a data-frame type class together. Specifically, this function implements an algorithm to bind resource tibbles. The function returns Joined (a tibble).
}
\keyword{internal}
