% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_recpts_tb}
\alias{print_recpts_tb}
\title{Print recpts}
\usage{
print_recpts_tb(
  recipients_tb,
  caption_1L_chr = "Player groups of interest",
  mkdn_tbl_ref_1L_chr = "tab:pgps",
  output_type_1L_chr = "HTML",
  use_rdocx_1L_lgl = T
)
}
\arguments{
\item{recipients_tb}{Recipients (a tibble)}

\item{caption_1L_chr}{Caption (a character vector of length one), Default: 'Player groups of interest'}

\item{mkdn_tbl_ref_1L_chr}{Mkdn table reference (a character vector of length one), Default: 'tab:pgps'}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'HTML'}

\item{use_rdocx_1L_lgl}{Use rdocx (a logical vector of length one), Default: T}
}
\description{
print_recpts_tb() is a Print function that prints output to console Specifically, this function implements an algorithm to print recpts tibble. The function is called for its side effects and does not return a value.
}
\keyword{internal}
