% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_resc_occupcy_tb}
\alias{transform_resc_occupcy_tb}
\title{Transform resource occupancy tibble}
\usage{
transform_resc_occupcy_tb(resc_occupcy_tb, resources_tb)
}
\arguments{
\item{resc_occupcy_tb}{Resource occupancy (a tibble)}

\item{resources_tb}{Resources (a tibble)}
}
\value{
Resource occupancy (a tibble)
}
\description{
transform_resc_occupcy_tb() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform resource occupancy tibble. The function returns Resource occupancy (a tibble).
}
\keyword{internal}
