% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_main_calcs_with_met_dmd}
\alias{update_main_calcs_with_met_dmd}
\title{Update main calculations with met demand}
\usage{
update_main_calcs_with_met_dmd(input_data_ls)
}
\arguments{
\item{input_data_ls}{Input data (a list)}
}
\value{
Input data (a list)
}
\description{
update_main_calcs_with_met_dmd() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update main calculations with met demand. The function returns Input data (a list).
}
\keyword{internal}
